auth-adapter/requirements.txt
flask==3.1.0
requests==2.32.3
PyJWT==2.10.1
gunicorn==23.0.0
